Refactoring ui tests scenarios.
