+++ app/src/androidTest/java/br/stone/mobiletraining/samilasantos/chucknorrisfacts/screenRandomFact/RandomFactActivityTest.kt
@@ -1,10 +1,11 @@
 package br.stone.mobiletraining.samilasantos.chucknorrisfacts.screenRandomFact
 
 import android.support.test.runner.AndroidJUnit4
-import br.stone.mobiletraining.samilasantos.chucknorrisfacts.screenRandomFact.RandomFactActivityScenarios.launchActivity
-import br.stone.mobiletraining.samilasantos.chucknorrisfacts.screenRandomFact.RandomFactActivityScenarios.runWithError
-import br.stone.mobiletraining.samilasantos.chucknorrisfacts.screenRandomFact.RandomFactActivityScenarios.runWithLongTimeDurationRequest
-import br.stone.mobiletraining.samilasantos.chucknorrisfacts.screenRandomFact.RandomFactActivityScenarios.runWithSuccess
+import br.stone.mobiletraining.samilasantos.chucknorrisfacts.common.Scenarios
+import br.stone.mobiletraining.samilasantos.chucknorrisfacts.common.Scenarios.runWithError
+import br.stone.mobiletraining.samilasantos.chucknorrisfacts.common.Scenarios.runWithNoNetwork
+import br.stone.mobiletraining.samilasantos.chucknorrisfacts.common.Scenarios.runWithSuccess
+import br.stone.mobiletraining.samilasantos.chucknorrisfacts.screenRandomFact.RandomFactActivityRobot.launchActivity
 import org.junit.Test
 import org.junit.runner.RunWith
 
@@ -24,7 +25,7 @@ class RandomFactActivityTest {
 
     @Test
     fun given_a_timeout_error_then_verify_if_timeout_message_shown() {
-        runWithLongTimeDurationRequest {
+        Scenarios.runWithLongTimeDurationRequest {
             launchActivity {
                 clickOnUpdateButton()
                 verifyIfTextMatches(RandomFactActivityMother.timeoutFeedback())
@@ -62,6 +63,17 @@ class RandomFactActivityTest {
                 clickOnUpdateButton()
                 isFactVisible()
                 verifyIfTextMatches(RandomFactActivityMother.fact)
+            }
+        }
+    }
+
+    @Test
+    fun once_has_no_network_then_verify_if_no_network_exception_dialog_is_shown() {
+        runWithNoNetwork {
+            launchActivity {
+                clickOnUpdateButton()
+                verifyIfTextMatches(RandomFactActivityMother.noNetworkFeedback())
+                closeDialog()
             }
         }
     }
