+++ app/src/androidTest/java/br/stone/mobiletraining/samilasantos/chucknorrisfacts/screenSearchFact/SearchFactActivityRobot.kt
@@ -7,15 +7,29 @@ import android.support.test.espresso.assertion.ViewAssertions.matches
 import android.support.test.espresso.matcher.ViewMatchers.isDisplayed
 import android.support.test.espresso.matcher.ViewMatchers.withId
 import android.support.test.espresso.matcher.ViewMatchers.withText
+import android.support.test.rule.ActivityTestRule
 import br.stone.mobiletraining.samilasantos.chucknorrisfacts.R
 import com.schibsted.spain.barista.assertion.BaristaVisibilityAssertions.assertDisplayed
 import com.schibsted.spain.barista.interaction.BaristaClickInteractions.clickOn
 import com.schibsted.spain.barista.interaction.BaristaListInteractions.scrollListToPosition
+import org.junit.Rule
 
 object SearchFactActivityRobot {
 
     private const val SETTINGS = "settings"
     private const val RETRY = "Retry"
+
+    @Rule
+    @JvmField
+    val activityRule =
+        ActivityTestRule<SearchFactActivity>(
+            SearchFactActivity::class.java, false, false
+        )
+
+    fun launchActivity(func: SearchFactActivityRobot.() -> Unit) {
+        activityRule.launchActivity(null)
+        func.invoke(this)
+    }
 
     fun type(word: String) {
         onView(withId(R.id.editSearchQuery)).perform(
