+++ app/src/androidTest/java/br/stone/mobiletraining/samilasantos/chucknorrisfacts/screenSearchFact/SearchFactActivityTest.kt
@@ -1,6 +1,11 @@
 package br.stone.mobiletraining.samilasantos.chucknorrisfacts.screenSearchFact
 
 import android.support.test.espresso.intent.Intents
+import br.stone.mobiletraining.samilasantos.chucknorrisfacts.common.Scenarios.runWithError
+import br.stone.mobiletraining.samilasantos.chucknorrisfacts.common.Scenarios.runWithLongTimeDurationRequest
+import br.stone.mobiletraining.samilasantos.chucknorrisfacts.common.Scenarios.runWithNoNetwork
+import br.stone.mobiletraining.samilasantos.chucknorrisfacts.common.Scenarios.runWithSuccess
+import br.stone.mobiletraining.samilasantos.chucknorrisfacts.screenSearchFact.SearchFactActivityRobot.launchActivity
 import org.junit.After
 import org.junit.Before
 import org.junit.Test
@@ -14,7 +19,7 @@ class SearchFactActivityTest {
 
     @Test
     fun once_activity_has_launched_should_present_the_initial_state() {
-        SearchFactActivityScenarios.launchActivity {
+        launchActivity {
             checkIfSearchBarIsVisible()
             checkIfInitialMessageIsVisible()
         }
@@ -22,8 +27,8 @@ class SearchFactActivityTest {
 
     @Test
     fun given_a_timeout_error_then_verify_if_timeout_message_shown() {
-        SearchFactActivityScenarios.runWithLongTimeDurationRequest {
-            SearchFactActivityScenarios.launchActivity {
+        runWithLongTimeDurationRequest {
+            launchActivity {
                 type("dev")
                 clickOnSearchButton()
                 checkIfErrorLayoutIsVisible()
@@ -36,8 +41,8 @@ class SearchFactActivityTest {
 
     @Test
     fun given_a_sever_error_then_verify_if_unavailable_provider_message_is_shown() {
-        SearchFactActivityScenarios.runWithError(500) {
-            SearchFactActivityScenarios.launchActivity {
+        runWithError(500) {
+            launchActivity {
                 type("dev")
                 clickOnSearchButton()
                 checkIfErrorLayoutIsVisible()
@@ -48,8 +53,8 @@ class SearchFactActivityTest {
 
     @Test
     fun given_a_malformed_body_then_verify_if_unexpected_data_message_is_shown() {
-        SearchFactActivityScenarios.runWithSuccess(SearchFactActivityMother.malFormedBody) {
-            SearchFactActivityScenarios.launchActivity {
+        runWithSuccess(SearchFactActivityMother.malFormedBody) {
+            launchActivity {
                 type("dev")
                 clickOnSearchButton()
                 checkIfErrorLayoutIsVisible()
@@ -60,8 +65,8 @@ class SearchFactActivityTest {
 
     @Test
     fun given_a_success_scenario_then_verify_if_success_state_is_shown() {
-        SearchFactActivityScenarios.runWithSuccess(SearchFactActivityMother.successBody) {
-            SearchFactActivityScenarios.launchActivity {
+        runWithSuccess(SearchFactActivityMother.successBody) {
+            launchActivity {
                 type("dev")
                 clickOnSearchButton()
                 checkIfListIsVisible()
@@ -71,8 +76,8 @@ class SearchFactActivityTest {
 
     @Test
     fun given_an_empty_result_then_verify_if_error_view_is_shown() {
-        SearchFactActivityScenarios.runWithSuccess(SearchFactActivityMother.emptyBody) {
-            SearchFactActivityScenarios.launchActivity {
+        runWithSuccess(SearchFactActivityMother.emptyBody) {
+            launchActivity {
                 type("dev")
                 clickOnSearchButton()
                 checkIfErrorLayoutIsVisible()
@@ -83,8 +88,8 @@ class SearchFactActivityTest {
 
     @Test
     fun given_a_not_found_result_then_verify_if_error_view_is_shown() {
-        SearchFactActivityScenarios.runWithError(404) {
-            SearchFactActivityScenarios.launchActivity {
+        runWithError(404) {
+            launchActivity {
                 type("dev")
                 clickOnSearchButton()
                 checkIfErrorLayoutIsVisible()
@@ -95,8 +100,8 @@ class SearchFactActivityTest {
 
     @Test
     fun given_a_bad_request_result_then_verify_if_error_view_is_shown() {
-        SearchFactActivityScenarios.runWithError(400) {
-            SearchFactActivityScenarios.launchActivity {
+        runWithError(400) {
+            launchActivity {
                 type("dev")
                 clickOnSearchButton()
                 checkIfErrorLayoutIsVisible()
@@ -107,8 +112,8 @@ class SearchFactActivityTest {
 
     @Test
     fun once_has_no_network_then_verify_if_error_view_is_shown() {
-        SearchFactActivityScenarios.runWithNoNetwork {
-            SearchFactActivityScenarios.launchActivity {
+        runWithNoNetwork {
+            launchActivity {
                 type("dev")
                 clickOnSearchButton()
                 checkIfErrorLayoutIsVisible()
